Move envelope to own file.  Improve docs.
